+++ src/Utils/Utils.hpp
@@ -92,7 +92,7 @@ namespace Dungeon {
         {
             bool cap = true;
 
-            for (int i = 0; i <= s.length(); i++) {
+            for (size_t i = 0; i <= s.length(); i++) {
                 if (isalpha(s[i]) && cap) {
                     s[i] = toupper(s[i]);
                     cap = false;
@@ -108,7 +108,7 @@ namespace Dungeon {
 
         static inline std::string& decapitalize(std::string&& s)
         {
-            for (int i = 0; i <= s.length(); i++) {
+            for (size_t i = 0; i <= s.length(); i++) {
                 if (isalpha(s[i])) {
                     s[i] = tolower(s[i]);
                 }
